+++ packages/react-native/src/components/SectionHeader/index.ts
@@ -2,3 +2,4 @@ export { default as SectionHeader } from './SectionHeader';
 export { default as SectionHeaderHeading } from './SectionHeaderHeading';
 export { default as SectionHeaderHelperText } from './SectionHeaderHelperText';
 export { default as SectionHeaderTextContent } from './SectionHeaderTextContent';
+export { default as SectionHeaderTrailingContent } from './SectionHeaderTrailingContent';