+++ packages/react-native/src/tokens/components/dark/illustrations.ts
@@ -3,5 +3,6 @@
  */
 
 export default {
+  colorMode: 'dark',
   mode: 'dark',
 } as const;