+++ packages/react-native/src/tokens/components/dark/toast.ts
@@ -5,6 +5,9 @@
 export default {
   backgroundColor: '#3f3f3f',
   borderRadius: 8,
-  gapHorizontal: 8,
+  gapHorizontal: 16,
   padding: 14,
+  stack: {
+    gapHorizontal: 8,
+  },
 } as const;